🚚 Put queries in common
