+++ frontend/app/lib/queries/index.ts
@@ -2,14 +2,12 @@ export * from "./compose-stacks";
 export * from "./deployments";
 export * from "./environments";
 export * from "./git-apps";
-export * from "./logs";
 export * from "./misc";
-export * from "./preview-templates";
 export * from "./projects";
 export * from "./server";
+export * from "./server-admin";
 export * from "./services";
 export * from "./shared";
-export * from "./server-admin";
 export * from "./templates";
-export * from "./users";
+export * from "./user";
 export * from "./workspaces";
